+++ greenfield/eldenring/data.py
@@ -5256,4 +5256,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:9a95a94a640b533050b78d694b549f1b78ea205291995b86bf60e8d3883d0f8f', 'module': 'data.py', 'body_sha256': 'sha256:4cff56db1d20320903d73a63ce4415b59faf0d4fc75d108cef9cd0a73e809653'}
+_GEN_STAMP = {'inputs_hash': 'sha256:4494cd8b28bdda236b0755ace00b47a999b2baf76a2b5e0c4227f8f22142fb1c', 'module': 'data.py', 'body_sha256': 'sha256:4cff56db1d20320903d73a63ce4415b59faf0d4fc75d108cef9cd0a73e809653'}